Added XPAth for Overstock
 and moved regex into separate file
